# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation BileMo",
        "description": "Documentation for BileMo project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Provides new token with basic authentication",
                "operationId": "App\\Controller\\UserController::login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/loginUser"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/loggedUser"
                    },
                    "400": {
                        "$ref": "#/components/responses/badRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/invalidCredential"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show details for the current logged user",
                "operationId": "App\\Controller\\UserController::show",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/loggedUser"
                    },
                    "401": {
                        "$ref": "#/components/responses/invalidToken"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "customerGlobal": {
                "description": "Customer's global properties",
                "properties": {
                    "id": {
                        "description": "General information",
                        "type": "integer"
                    },
                    "firstname": {
                        "description": "General information",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "General information",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "userGlobal": {
                "description": "User's global properties",
                "properties": {
                    "id": {
                        "description": "General information",
                        "type": "integer"
                    },
                    "firstname": {
                        "description": "General information",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "General information",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "Customer's properties",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/userGlobal"
                    },
                    {
                        "properties": {
                            "email": {
                                "description": "General information",
                                "type": "string"
                            },
                            "": {
                                "description": "General information",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/customerGlobal"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "responses": {
...
            "loggedUser": {
                "description": "return details from logged user",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                }
            },
...
            "invalidToken": {
                "description": "return an error message according to an invalid, missing or expired token",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "General information",
                                    "type": "integer",
                                    "example": "401"
                                },
                                "message": {
                                    "description": "General information",
                                    "type": "string",
                                    "example": "invalid token"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "invalidCredential": {
                "description": "return an invalid credential error message",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "General information",
                                    "type": "integer",
                                    "example": "401"
                                },
                                "message": {
                                    "description": "General information",
                                    "type": "string",
                                    "example": "invalid credentials"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "badRequest": {
                "description": "return a message with the invalid data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "General information",
                                    "type": "integer",
                                    "example": "400"
                                },
                                "message": {
                                    "description": "General information",
                                    "type": "string",
                                    "example": "This Json content invalid data."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "requestBodies": {
...
            "loginUser": {
                "description": "General information",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "username": {
                                    "description": "General information",
                                    "type": "string",
                                    "example": "username@email.com"
                                },
                                "password": {
                                    "description": "General information",
                                    "type": "string",
                                    "example": "MySecretPassword"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}